+++ api/src/main/kotlin/com/ford/internalprojects/peoplemover/user/exceptions/InvalidUserModification.kt
@@ -1,4 +1,4 @@
-package com.ford.internalprojects.peoplemover.space.exceptions
+package com.ford.internalprojects.peoplemover.user.exceptions
 
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
