+++ api/src/test/kotlin/com/ford/internalprojects/peoplemover/space/SpaceControllerApiTest.kt
@@ -228,223 +228,7 @@ class SpaceControllerApiTest {
             .andExpect(status().isBadRequest)
     }
 
-    @Test
-    fun `GET should return all user id that have edit access for a space`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-        val space2: Space = spaceRepository.save(Space(name = "SpaceTwo"))
 
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "editor"))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "ANOTHER_USER_ID", spaceUuid = space1.uuid, permission = "editor"))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "ANOTHER_SPACE_ANOTHER_USER_ID", spaceUuid = space2.uuid, permission = "editor"))
-
-        val result = mockMvc.perform(
-            get(baseSpaceUrl + "/${space1.uuid}/editors")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-        )
-        .andExpect(status().isOk)
-        .andReturn()
-
-        val returnedEditors: List<String> = objectMapper.readValue(
-            result.response.contentAsString,
-            objectMapper.typeFactory.constructCollectionType(MutableList::class.java, String::class.java)
-        )
-
-        assertThat(returnedEditors).containsExactlyInAnyOrder("USER_ID", "ANOTHER_USER_ID")
-    }
-
-    @Test
-    fun `GET should return all user id that have edit access for a space and ignore empty and null user ids`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "editor"))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "ANOTHER_USER_ID", spaceUuid = space1.uuid, permission = "editor"))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "", spaceUuid = space1.uuid, permission = "editor"))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = null, spaceUuid = space1.uuid, permission = "editor"))
-
-        val result = mockMvc.perform(
-                get(baseSpaceUrl + "/${space1.uuid}/editors")
-                        .header("Authorization", "Bearer GOOD_TOKEN")
-        )
-                .andExpect(status().isOk)
-                .andReturn()
-
-        val returnedEditors: List<String> = objectMapper.readValue(
-                result.response.contentAsString,
-                objectMapper.typeFactory.constructCollectionType(MutableList::class.java, String::class.java)
-        )
-
-        assertThat(returnedEditors).containsExactlyInAnyOrder("USER_ID", "ANOTHER_USER_ID", "UNKNOWN_USER", "UNKNOWN_USER")
-    }
-
-
-    @Test
-    fun `GET should return 403 if the user does not have write access to the space`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "editor"))
-
-        mockMvc.perform(
-            get(baseSpaceUrl + "/${space1.uuid}/editors")
-                .header("Authorization", "Bearer ANONYMOUS_TOKEN")
-        )
-            .andExpect(status().isForbidden)
-            .andReturn()
-    }
-
-    @Test
-    fun `GET should return 403 if the space does not exist`() {
-        mockMvc.perform(
-            get(baseSpaceUrl + "/aaaaaaaa-aaaa-aaaa-aaaa-badspace1234/editors")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-        )
-            .andExpect(status().isForbidden)
-            .andReturn()
-    }
-
-    @Test
-    fun `GET should return all users for a space`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-
-        val user1 = UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "owner")
-        val user2 = UserSpaceMapping(userId = "ANOTHER_USER_ID", spaceUuid = space1.uuid, permission = "editor")
-        val user3 = UserSpaceMapping(userId = "", spaceUuid = space1.uuid, permission = "editor")
-        val user4 = UserSpaceMapping(userId = null, spaceUuid = space1.uuid, permission = "editor")
-        userSpaceMappingRepository.save(user1)
-        userSpaceMappingRepository.save(user2)
-        userSpaceMappingRepository.save(user3)
-        userSpaceMappingRepository.save(user4)
-
-        val result = mockMvc.perform(
-            get(baseSpaceUrl + "/${space1.uuid}/users")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-        )
-            .andExpect(status().isOk)
-            .andReturn()
-
-        val returnedEditors: List<UserSpaceMapping> = objectMapper.readValue(
-            result.response.contentAsString,
-            objectMapper.typeFactory.constructCollectionType(MutableList::class.java, UserSpaceMapping::class.java)
-        )
-
-        assertThat(returnedEditors).hasSize(4)
-        assertThat(returnedEditors).contains(user1);
-        assertThat(returnedEditors).contains(user2);
-        assertThat(returnedEditors).contains(user3);
-        assertThat(returnedEditors).contains(user4);
-
-    }
-
-    @Test
-    fun `DELETE should remove user for a space`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-
-        val user1 = UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "owner")
-        val user2 = UserSpaceMapping(userId = "ANOTHER_USER_ID", spaceUuid = space1.uuid, permission = "editor")
-        userSpaceMappingRepository.save(user1)
-        userSpaceMappingRepository.save(user2)
-
-        mockMvc.perform(
-            delete(baseSpaceUrl + "/${space1.uuid}/users/${user2.userId}")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-        )
-            .andExpect(status().isOk)
-
-        val users = userSpaceMappingRepository.findAllBySpaceUuid(space1.uuid)
-
-        assertThat(users).hasSize(1)
-        assertThat(users).contains(user1);
-        assertThat(users).doesNotContain(user2);
-    }
-
-    @Test
-    fun `DELETE should return 403 if the user does not have write access to the space`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-
-        val user1 = UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "owner")
-        val user2 = UserSpaceMapping(userId = "ANOTHER_USER_ID", spaceUuid = space1.uuid, permission = "editor")
-        userSpaceMappingRepository.save(user1)
-        userSpaceMappingRepository.save(user2)
-
-        mockMvc.perform(
-                delete(baseSpaceUrl + "/${space1.uuid}/users/${user2.userId}")
-                        .header("Authorization", "Bearer ANONYMOUS_TOKEN")
-        )
-        .andExpect(status().isForbidden)
-        .andReturn()
-
-        val users = userSpaceMappingRepository.findAllBySpaceUuid(space1.uuid)
-
-        assertThat(users).hasSize(2)
-        assertThat(users).contains(user1);
-        assertThat(users).contains(user2);
-    }
-
-    @Test
-    fun `DELETE should return 406 if the caller tries to delete owner`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-
-        val user1 = UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "owner")
-        val user2 = UserSpaceMapping(userId = "ANOTHER_USER_ID", spaceUuid = space1.uuid, permission = "editor")
-        userSpaceMappingRepository.save(user1)
-        userSpaceMappingRepository.save(user2)
-
-        mockMvc.perform(
-                delete(baseSpaceUrl + "/${space1.uuid}/users/${user1.userId}")
-                    .header("Authorization", "Bearer GOOD_TOKEN")
-        )
-        .andExpect(status().isNotAcceptable)
-        .andReturn()
-
-        val users = userSpaceMappingRepository.findAllBySpaceUuid(space1.uuid)
-
-        assertThat(users).hasSize(2)
-        assertThat(users).contains(user1);
-        assertThat(users).contains(user2);
-    }
-
-    @Test
-    fun `DELETE should return 400 if the user doesn't exist`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-
-        val user1 = UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "owner")
-        val user2 = UserSpaceMapping(userId = "ANOTHER_USER_ID", spaceUuid = space1.uuid, permission = "editor")
-        userSpaceMappingRepository.save(user1)
-
-        mockMvc.perform(
-                delete(baseSpaceUrl + "/${space1.uuid}/users/${user2.userId}")
-                    .header("Authorization", "Bearer GOOD_TOKEN")
-        )
-        .andExpect(status().isBadRequest)
-        .andReturn()
-
-        val users = userSpaceMappingRepository.findAllBySpaceUuid(space1.uuid)
-
-        assertThat(users).hasSize(1)
-        assertThat(users).contains(user1);
-    }
-
-    @Test
-    fun `DELETE should return 400 if trying to delete a user from another space`() {
-        val space1: Space = spaceRepository.save(Space(name = "SpaceOne"))
-        val space2: Space = spaceRepository.save(Space(name = "SpaceTwo"))
-
-        val user1 = UserSpaceMapping(userId = "USER_ID", spaceUuid = space1.uuid, permission = "owner")
-        val user2 = UserSpaceMapping(userId = "ANOTHER_USER_ID", spaceUuid = space2.uuid, permission = "editor")
-        userSpaceMappingRepository.save(user1)
-        userSpaceMappingRepository.save(user2)
-
-        mockMvc.perform(
-            delete(baseSpaceUrl + "/${space1.uuid}/users/${user2.userId}")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-        )
-            .andExpect(status().isBadRequest)
-            .andReturn()
-
-        val users = userSpaceMappingRepository.findAllBySpaceUuid(space2.uuid)
-
-        assertThat(users).hasSize(1)
-        assertThat(users).contains(user2);
-    }
 
     @Test
     fun `Edit Space Request should return 200 if space is edited correctly`() {
@@ -597,130 +381,4 @@ class SpaceControllerApiTest {
         assertThat(actualSpace!!.name).isEqualTo(expectedName)
     }
 
-    @Test
-    fun `Invite Users Request should return Ok and an empty list with a valid ADFS request`() {
-        val emails = listOf("email1", "email2")
-
-        val space = spaceRepository.save(Space(name = "spaceName"))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "USER_ID", spaceUuid = space.uuid, permission = PERMISSION_OWNER))
-
-        val request = AuthInviteUsersToSpaceRequest(userIds = emails)
-
-        val result = mockMvc.perform(
-            put("$baseSpaceUrl/${space.uuid}:invite")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-                .content(objectMapper.writeValueAsString(request))
-                .contentType("application/json")
-        ).andExpect(
-            status().isOk
-        ).andReturn()
-
-        assertThat(result.response.contentLength).isEqualTo(0)
-
-        val savedIds = userSpaceMappingRepository.findAll().map { Pair(it.userId!!, it.permission)}
-
-        assertThat(userSpaceMappingRepository.count()).isEqualTo(3)
-        assertThat(savedIds).contains(Pair("USER_ID", PERMISSION_OWNER))
-        assertThat(savedIds).contains(Pair("EMAIL1", PERMISSION_EDITOR))
-        assertThat(savedIds).contains(Pair("EMAIL2", PERMISSION_EDITOR))
-    }
-
-    @Test
-    fun `Invite Users Request should return BAD_REQUEST if no emails were provided`() {
-        val request = AuthInviteUsersToSpaceRequest(
-            userIds = listOf()
-        )
-
-        val space = spaceRepository.save(Space(name = "spaceName"))
-
-        mockMvc.perform(
-            put("$baseSpaceUrl/${space.uuid}:invite")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-                .content(objectMapper.writeValueAsString(request))
-                .contentType("application/json")
-        ).andExpect(
-            status().isBadRequest
-        )
-    }
-    @Test
-    fun `Invite Users Request should return BAD_REQUEST if cdsids are not valid`() {
-        val request = AuthInviteUsersToSpaceRequest(
-            userIds = listOf("abc12", "1@5")
-        )
-
-        val space = spaceRepository.save(Space(name = "spaceName"))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "USER_ID", spaceUuid = space.uuid, permission = PERMISSION_OWNER))
-
-        mockMvc.perform(
-            put("$baseSpaceUrl/${space.uuid}:invite")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-                .content(objectMapper.writeValueAsString(request))
-                .contentType("application/json")
-        ).andExpect(
-            status().isBadRequest
-        )
-    }
-
-    @Test
-    fun `Invite Users Request should return 403 when trying to add users to space without write authorization`() {
-        val space = spaceRepository.save(Space(name = "spaceName"))
-
-        val requestBodyObject =
-            AuthInviteUsersToSpaceRequest(userIds = listOf("email1", "email2"))
-
-        mockMvc.perform(
-            put("$baseSpaceUrl/${space.uuid}:invite")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(requestBodyObject))
-        )
-            .andExpect(status().isForbidden)
-    }
-
-    @Test
-    fun `Set owner request should return 403 when trying to set owner with editor permissions`() {
-        val space = spaceRepository.save(Space(name = "spaceName"))
-
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "USER_ID", spaceUuid = space.uuid, permission = PERMISSION_EDITOR))
-
-        mockMvc.perform(
-            put("$baseSpaceUrl/${space.uuid}/users/USER_ID")
-                .header("Authorization", "Bearer GOOD_TOKEN")
-                .contentType(MediaType.APPLICATION_JSON)
-        ).andExpect(status().isForbidden)
-    }
-
-    @Test
-    fun `Set owner request should return 200 when owner tries to assign ownership to an editor`() {
-        val space = spaceRepository.save(Space(name = "spaceName"))
-
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "EDITOR_ID", spaceUuid = space.uuid, permission = PERMISSION_EDITOR))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "USER_ID", spaceUuid = space.uuid, permission = PERMISSION_OWNER))
-
-        mockMvc.perform(
-                put("$baseSpaceUrl/${space.uuid}/users/EDITOR_ID")
-                        .header("Authorization", "Bearer GOOD_TOKEN")
-                        .contentType(MediaType.APPLICATION_JSON)
-        ).andExpect(status().isOk)
-
-        assertThat(userSpaceMappingRepository.findAllByUserId("EDITOR_ID")[0].permission).isEqualTo(PERMISSION_OWNER)
-        assertThat(userSpaceMappingRepository.findAllByUserId("USER_ID")[0].permission).isEqualTo(PERMISSION_EDITOR)
-    }
-
-    @Test
-    fun `Set owner request should return 400 when owner tries to assign ownership to an editor that does not exist`() {
-        val space = spaceRepository.save(Space(name = "spaceName"))
-
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "EDITOR_ID", spaceUuid = space.uuid, permission = PERMISSION_EDITOR))
-        userSpaceMappingRepository.save(UserSpaceMapping(userId = "USER_ID", spaceUuid = space.uuid, permission = PERMISSION_OWNER))
-
-        mockMvc.perform(
-                put("$baseSpaceUrl/${space.uuid}/users/INVALID_ID")
-                        .header("Authorization", "Bearer GOOD_TOKEN")
-                        .contentType(MediaType.APPLICATION_JSON)
-        ).andExpect(status().isBadRequest)
-
-        assertThat(userSpaceMappingRepository.findAllByUserId("EDITOR_ID")[0].permission).isEqualTo(PERMISSION_EDITOR)
-        assertThat(userSpaceMappingRepository.findAllByUserId("USER_ID")[0].permission).isEqualTo(PERMISSION_OWNER)
-    }
 }
